requirements.txt
autopep8==1.4.4
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.0
pycodestyle==2.5.0
pycparser==2.19
pyparsing==2.4.2
six==1.12.0
SQLAlchemy==1.3.9
svgwrite==1.3.1
Werkzeug==0.16.0
WTForms==2.2.1
